requirements.txt
feedparser
requests
python-dotenv
google-generativeai


